# Conflicts:
#	coko/com.elster.jupiter.issue.extjs/src/main/resources/i18n.properties
#	coko/com.elster.jupiter.webservices.inbound.rest.demo/pom.xml
#	comu/com.energyict.mdc.firmware.extjs/src/main/resources/i18n.properties
#	copl/com.elster.jupiter.http.whiteboard/src/main/java/com/elster/jupiter/http/whiteboard/impl/BasicAuthentication.java
#	copl/com.elster.jupiter.http.whiteboard/src/main/java/com/elster/jupiter/http/whiteboard/impl/WhiteBoardImpl.java
#	copl/com.elster.jupiter.orm/src/main/java/com/elster/jupiter/orm/Table.java

+++ coko/com.elster.jupiter.export/src/main/java/com/elster/jupiter/export/impl/UpgraderV10_7_2.java
@@ -3,6 +3,7 @@
  */
 package com.elster.jupiter.export.impl;
 
+import com.elster.jupiter.export.impl.webservicecall.DataExportServiceCallTypeImpl;
 import com.elster.jupiter.orm.DataModel;
 import com.elster.jupiter.orm.DataModelUpgrader;
 import com.elster.jupiter.orm.Version;
@@ -12,18 +13,23 @@ import com.google.inject.Inject;
 
 public class UpgraderV10_7_2 implements Upgrader {
     private final DataModel dataModel;
+    private final DataExportServiceCallTypeImpl dataExportServiceCallType;
 
     @Inject
-    UpgraderV10_7_2(DataModel dataModel) {
+    UpgraderV10_7_2(DataModel dataModel, DataExportServiceCallTypeImpl dataExportServiceCallType) {
         this.dataModel = dataModel;
+        this.dataExportServiceCallType = dataExportServiceCallType;
     }
 
     @Override
     public void migrate(DataModelUpgrader dataModelUpgrader) {
         dataModelUpgrader.upgrade(dataModel, Version.version(10, 7, 2));
-        execute(dataModel,
-                "update DES_RTDATAEXPORTITEM set READING_INTERVAL = '1 hours' where READINGINTERVALCOUNT is not null and READINGINTERVALCOUNT != 0",
-                "alter table DES_RTDATAEXPORTITEM drop column READINGINTERVALCOUNT",
-                "alter table DES_RTDATAEXPORTITEM drop column READINGINTERVALUNIT");
+        if (dataModel.doesColumnExist("DES_RTDATAEXPORTITEM", "READINGINTERVALCOUNT")) {
+            execute(dataModel,
+                    "update DES_RTDATAEXPORTITEM set READING_INTERVAL = '1 hours' where READINGINTERVALCOUNT is not null and READINGINTERVALCOUNT != 0",
+                    "alter table DES_RTDATAEXPORTITEM drop column READINGINTERVALCOUNT",
+                    "alter table DES_RTDATAEXPORTITEM drop column READINGINTERVALUNIT");
+        }
+        dataExportServiceCallType.findOrCreateChildType();
     }
 }
